+++ src/DpressServices.php
@@ -30,6 +30,7 @@ use Dynart\Micro\Entities\QueryExecutor;
 use Dynart\Dpress\Cli\SchemaCommands;
 use Dynart\Dpress\Cli\SystemCommands;
 use Dynart\Dpress\Cli\UserCommands;
+use Dynart\Dpress\Content\Callouts;
 use Dynart\Dpress\Content\CodeAssets;
 use Dynart\Dpress\Content\CodeBlockRenderer;
 use Dynart\Dpress\Content\InternalLinks;
@@ -193,6 +194,7 @@ class DpressServices {
         Micro::add(LinkTargetResolverInterface::class, LinkTargets::class);
         Micro::add(InternalLinks::class);
         Micro::add(Shortcodes::class);
+        Micro::add(Callouts::class);
         Micro::add(CodeAssets::class);
         Micro::add(CodeBlockRenderer::class);
         Micro::add(ShortcodeRenderer::class);
@@ -238,6 +240,7 @@ class DpressServices {
         $events->subscribe(MarkdownRenderer::EVENT_ENVIRONMENT, [InternalLinks::class, 'onEnvironment']);
         $events->subscribe(MarkdownRenderer::EVENT_ENVIRONMENT, [ShortcodeRenderer::class, 'onEnvironment']);
         $events->subscribe(MarkdownRenderer::EVENT_ENVIRONMENT, [CodeBlockRenderer::class, 'onEnvironment']);
+        $events->subscribe(MarkdownRenderer::EVENT_ENVIRONMENT, [Callouts::class, 'onEnvironment']);
     }
 
     /**